[Feat] user模块业务完善

+++ common/src/main/java/org/zhaoxuan/common/constants/TimeConstants.java
@@ -3,11 +3,11 @@ package org.zhaoxuan.common.constants;
 @SuppressWarnings("unused")
 public interface TimeConstants {
 
-    long ONE_SECOND = 1000;
-    long ONE_MINUTE = ONE_SECOND * 60;
-    long HALF_HOUR = ONE_MINUTE * 30;
-    long ONE_HOUR = HALF_HOUR * 2;
-    long HALF_DAY = ONE_HOUR * 12;
-    long ONE_DAY = HALF_DAY * 2;
+    int ONE_SECOND = 1000;
+    int ONE_MINUTE = ONE_SECOND * 60;
+    int HALF_HOUR = ONE_MINUTE * 30;
+    int ONE_HOUR = HALF_HOUR * 2;
+    int HALF_DAY = ONE_HOUR * 12;
+    int ONE_DAY = HALF_DAY * 2;
 
 }
